XPages security fixes, doc sign/encrypt, ID Vault APIs, SSO tokens
Major rewrite and class refactoring to prevent SecurityManager
exceptions when using the API in an XPages application.
JNA structures classes for Notes datatypes now end with "...Struct" and
provide a newInstance() method that we use instead of the normal
constructor, because we had to wrap class instantiation in
AccessController.doPrivileged blocks. Old constructors had to be kept
because JNA uses them for structure size computation.
For most use cases, those structures should not be needed anyway: We
started to introduce wrapper classes like NotesTimeDate that are not
subclasses of JNA structures anymore.
This commit also contains APIs to download ID files from the ID vault
(either to disk or kept in-memory), sign, encrypt and decrypt documents
and a new utility class to compute SSO tokens using the Domino server
APIs.