Went back to 0.9.4, fixing JNA issues and GLIBC system requirements
Set jna.nosys=true to enforce using our extracted JNA dll/so file and
not another version we find on the PATH. Upgraded to JNA 4.4, but
switched to the org.elasticsearch version. Elasticsearch recompiled the
JNA Linux binary to not require GLIBC 2.14 to be installed on Linux
systems like standard JNA (see elasticsearch#23640).