Fixed duplicate handle warning, new method for collection indexer data
Duplicate handle warning was thrown when using
LegacyUtils.toNotesDatabase(Database) multiple times with the same
database. We know skip this check for API objects that return
isNoRecycle()==true.
The new API gives access to the total doc count and the percentile info
for the collection index (e.g. view data at 10%, 20%, ... 100% of the
view)