Session removed from NotesDatabase constructors; new name match function
We removed the lotus.domino.Session argument from some of the
NotesDatabase constructors, because it is not needed anymore since we
added the C code to read the view design / column sorting in
NotesCollection.
NotesNamingUtils.nameMatchesWildcard(String name, String wildcard) is a
new function that compares Notes names against wildcards like */Company,
hopefully working the way Domino does it.