Use NSFBuildNamesList for NAMES_LIST creation on DB open
We don't write our own NAMES_LIST structure anymore in this simple case,
once again to reduce the risk of incompatibility and server crashes.
Since the DB is already opened as a specific user, we now use
NIFOpenCollection instead of NIFOpenCollectionWithUserNameList to open
views.
In our tests we verified that the returned view data is still filtered
correctly for the user opening the database when reader fields are in
use.