Added standard method to write name of effective user
We could not find a C API method for Session.getEffectiveUsername() yet,
which when running in a web app returns the name of the webuser
(IDUtils.getCurrentUsername() returns the ID file owner). So at least we
are adding a standard method that can be used to store the effective
username for the current NotesHC execution block.