Renamed method IDUtils.getCurrentUsername() to IDUtils.getIdUsername()
Want to make it more clear which name we actually return here,
especially when running in a web app where the server name is returned.
Use IDUtils.setEffectiveUsername(String) and getEffectiveUsername() to
store the name of the web user for the duration of the current NotesGC
block. There's currently no known way to get this effective username via
C call. IBM dev for the Java API said that the name for
Session.getEffectiveUsername() is handed over to them from the agent
manager or HTTP task and cached locally, just like we do here in Domino
JNA.